Change name to MUGS (MongoDB Using Google Sheets)
Rationale for changing name from MIGS (MongoDB in Google Sheets) to MUGS:

(1) It's more accurate to say "using" than to say "in". The MongoDB
database is not actually contained within a Google Sheets spreadsheet;
rather, the spreadsheet interacts with the database. In other words,
using the spreadsheet to do MongoDB things.

(2) MUGS is an English word, MIGS is not.

That's enough bikeshedding, we almost certainly won't change the name
again.